Fix image upload to correct path in Firebase
Had to add yet another context to the app because yeah
