+++ whiteboardflow-frontend/src/Main.js
@@ -11,6 +11,7 @@ import QuestionSelect from './components/QuestionSelect/QuestionSelect'
 import { auth } from './firebase';
 import { AvatarProvider } from './components/Avatar/AvatarContext';
 import { QuestionProvider } from './components/QuestionSelect/QuestionContext';
+import { WhiteboardProvider } from './components/Whiteboard/WhiteboardContext';
 import History from './components/History/History'
 import HistoryEntry from './components/History/HistoryEntry';
 
@@ -31,23 +32,25 @@ function Main() {
 
     return (
         <Router>
-            <AvatarProvider>
-                <QuestionProvider>
-                    <Routes>
-                        <Route element={<Layout user={user} />}>
-                            <Route index element={<HomePage user={user} />} />
-                            {/* Only allow access to these routes if user is logged in */}
-                            <Route path="OralTest" element={<PrivateRoute><OralTest /></PrivateRoute>} />
-                            <Route path="results" element={<PrivateRoute><Results /></PrivateRoute>} />
-                            <Route path="whiteboard" element={<PrivateRoute><Whiteboard /></PrivateRoute>} />
-                            <Route path="difficulty" element={<PrivateRoute><DifficultySelect /></PrivateRoute>} />
-                            <Route path="questionSelect" element={<PrivateRoute><QuestionSelect /></PrivateRoute>} />
-                            <Route path="history" element={<PrivateRoute><History /></PrivateRoute>} />
-                            <Route path="history/:entry_id" element={<PrivateRoute><HistoryEntry /></PrivateRoute>} />
-                        </Route>
-                    </Routes>
-                </QuestionProvider>
-            </AvatarProvider>
+            <WhiteboardProvider>
+                <AvatarProvider>
+                    <QuestionProvider>
+                        <Routes>
+                            <Route element={<Layout user={user} />}>
+                                <Route index element={<HomePage user={user} />} />
+                                {/* Only allow access to these routes if user is logged in */}
+                                <Route path="questionSelect" element={<PrivateRoute><QuestionSelect /></PrivateRoute>} />
+                                <Route path="whiteboard" element={<PrivateRoute><Whiteboard /></PrivateRoute>} />
+                                <Route path="OralTest" element={<PrivateRoute><OralTest /></PrivateRoute>} />
+                                <Route path="results" element={<PrivateRoute><Results /></PrivateRoute>} />
+                                <Route path="difficulty" element={<PrivateRoute><DifficultySelect /></PrivateRoute>} />
+                                <Route path="history" element={<PrivateRoute><History /></PrivateRoute>} />
+                                <Route path="history/:entry_id" element={<PrivateRoute><HistoryEntry /></PrivateRoute>} />
+                            </Route>
+                        </Routes>
+                    </QuestionProvider>
+                </AvatarProvider>
+            </WhiteboardProvider>
         </Router>
     );
 }
